refactor(storage_impl): Move Conversion impl for authentication to storage_impl crate
