+++ crates/storage_impl/src/lib.rs
@@ -10,6 +10,8 @@ use hyperswitch_domain_models::{
 use hyperswitch_masking::StrongSecret;
 use redis::{kv_store::RedisConnInterface, pub_sub::PubSubInterface, RedisStore};
 mod address;
+pub mod authentication;
+pub mod behaviour;
 pub mod business_profile;
 pub mod callback_mapper;
 pub mod card_issuer;
